requirements.txt
# Core dependencies
streamlit==1.31.0
python-dotenv==1.0.0

# LLM and Agent frameworks
langchain==0.1.9
langchain-google-genai==0.0.11
langgraph==0.0.26
google-generativeai==0.4.1

# Data processing
pandas==2.2.0
duckdb==0.10.0

# Vector database
chromadb==0.4.22

# Utilities
pydantic==2.6.1
pydantic-settings==2.1.0
